requirements.txt
# Core ML and Computer Vision
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Web Framework and API
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0.0
requests>=2.31.0

# Video Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
pillow>=10.0.0
av>=10.0.0

# Data Processing
pandas>=2.0.0
h5py>=3.9.0
joblib>=1.3.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Deployment
gunicorn>=21.0.0
docker>=6.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0.0
click>=8.1.0
